Geometry param inputs
