+++ Runtime/Module.Fur/FurSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: df4e250bac4be4b47b9772f9b07d8c7b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
